added cleanJson scripts to coreLabelTools.py added removeGeoJSONIdField.py script for command line use
